requirements.txt
pandas>=2.2.0
requests>=2.31.0
python-dateutil>=2.9.0
pydantic>=2.8.0
tenacity>=8.2.2
pyyaml
pyyaml
